requirements.txt
dj_database_url==0.3.0
django-annoying==0.8.0
django-cities-light==3.0.2
django-cors-headers==1.0.0
django-debug-toolbar
django-jenkins
django-widget-tweaks==1.3
django==1.7.1
django_filter==0.9.1
django_pipeline==1.4.2
djangorestframework==3.0.2
djorm-ext-pgjson==0.3
djorm-pgarray==1.2
git+https://github.com/disqus/django-bitfield.git
gunicorn==19.1.1
psycopg2==2.5.4
rest_framework_ember==1.1.0
stripe==1.20.2
